+++ src/async-queue/publish/async_queue.php
@@ -13,7 +13,9 @@ declare(strict_types=1);
 return [
     'default' => [
         'driver' => \Hyperf\AsyncQueue\Driver\RedisDriver::class,
-        'pool' => 'default',
+        'redis' => [
+            'pool' => 'default',
+        ],
         'channel' => 'queue',
         'timeout' => 2,
         'retry_seconds' => 5,