+++ src/async-queue/tests/DriverTest.php
@@ -14,6 +14,7 @@ namespace HyperfTest\AsyncQueue;
 
 use Hyperf\AsyncQueue\Driver\ChannelConfig;
 use Hyperf\Di\Container;
+use Hyperf\Redis\RedisFactory;
 use Hyperf\Utils\ApplicationContext;
 use Hyperf\Utils\Coroutine\Concurrent;
 use Hyperf\Utils\Packer\PhpSerializerPacker;
@@ -59,6 +60,11 @@ class DriverTest extends TestCase
         $container->shouldReceive('get')->with(EventDispatcherInterface::class)->andReturn(null);
         $container->shouldReceive('make')->with(ChannelConfig::class, Mockery::any())->andReturnUsing(function ($class, $args) {
             return new ChannelConfig($args['channel']);
+        });
+        $container->shouldReceive('get')->with(RedisFactory::class)->andReturnUsing(function ($_) {
+            $factory = Mockery::mock(RedisFactory::class);
+            $factory->shouldReceive('get')->with('default')->andReturn(new Redis());
+            return $factory;
         });
 
         ApplicationContext::setContainer($container);